+++ Coimbra.Tests/ActorTests.cs
@@ -76,7 +76,7 @@ namespace Coimbra.Tests
         {
             const string logFormat = "OnDestroyed.reason = {0}";
             Actor instance = new GameObject().AddComponent<Actor>();
-            instance.OnDestroyed += delegate(Actor sender, Actor.DestroyReason reason)
+            instance.OnDestroying += delegate(Actor sender, Actor.DestroyReason reason)
             {
                 Debug.LogFormat(logFormat, reason);
             };